backend/django/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
django-netfields==1.3.2
environs==11.0.0
psycopg[binary,pool]==3.2.2
boto3==1.35.3
pytest-django==4.9.0
drf-yasg
ipython
openfga_sdk
django-storages
djangorestframework
django-filter
django-import-export[all]
gunicorn
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-exporter-prometheus
prometheus_client